Rename files
Renamed files as per the task description
