+++ app/src/main/java/cs/dal/krush/studentFragments/StudentPaymentFragment.java
@@ -18,7 +18,6 @@ import android.widget.TextView;
 import java.text.NumberFormat;
 
 import cs.dal.krush.R;
-import cs.dal.krush.appFragments.SessionDetailsFragment;
 import cs.dal.krush.models.DBHelper;
 
 /**
@@ -180,7 +179,7 @@ public class StudentPaymentFragment extends Fragment {
                     bundle.putInt("SESSION_ID", SESSION_ID);
 
                     // Swap into new fragment
-                    SessionDetailsFragment session = new SessionDetailsFragment();
+                    StudentUpcSessionsDetailsFragment session = new StudentUpcSessionsDetailsFragment();
                     session.setArguments(bundle);
                     FragmentTransaction transaction = getFragmentManager().beginTransaction();
                     transaction.replace(R.id.student_fragment_container, session);
