+++ app/src/main/java/cs/dal/krush/tutorFragments/TutorHistoryFragment.java
@@ -1,4 +1,4 @@
-package cs.dal.krush.studentFragments;
+package cs.dal.krush.tutorFragments;
 
 import android.content.Context;
 import android.database.Cursor;
@@ -12,25 +12,25 @@ import android.view.ViewGroup;
 import android.widget.AdapterView;
 import android.widget.ListView;
 import android.widget.TextView;
-import cs.dal.krush.StudentCursorAdapters.SessionCursorAdapter;
+
 import cs.dal.krush.R;
+import cs.dal.krush.TutorCursorAdapters.SessionCursorAdapter;
 import cs.dal.krush.models.DBHelper;
 
 /**
- * Sets up the Student Sessions History fragment. This fragment belongs to the StudentMainActivity class
- * and is accessed through the student's bottom navigation bar.
+ * Sets up the Tutor Home fragment. This fragment belongs to the TutorMainActivity class
+ * and is accessed through the tutor's bottom navigation bar.
  *
- * The student can view their previous sessions history along with audio recordings using this fragment.
+ * The tutor can view their session history using this fragment.
  */
-
-public class StudentSessionsFragment extends Fragment {
+public class TutorHistoryFragment extends Fragment {
 
     static int USER_ID;
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.student_sessions, container, false);
+        View view = inflater.inflate(R.layout.tutor_sessions, container, false);
         USER_ID = getArguments().getInt("USER_ID");
 
         //get Context:
@@ -38,6 +38,9 @@ public class StudentSessionsFragment extends Fragment {
 
         //init DB connection:
         DBHelper mydb = new DBHelper(C);
+
+        Cursor tutor = mydb.tutor.getData(USER_ID);
+        tutor.moveToFirst();
 
         //fetch UI elements:
         ListView sessionHistoryListView = (ListView)view.findViewById(R.id.sessionHistoryListView);
@@ -49,8 +52,8 @@ public class StudentSessionsFragment extends Fragment {
         //set font style:
         pageTitle.setTypeface(typeFace);
 
-        //get all tutoring sessions by the student:
-        final Cursor cursorSessionsResponse = mydb.tutoringSession.getSessionHistoryByStudentIdForCursorAdapter(USER_ID);
+        //get all tutoring sessions by the tutor:
+        final Cursor cursorSessionsResponse = mydb.tutoringSession.getSessionHistoryByTutorIdForCursorAdapter(USER_ID);
 
         //set sessions listview adapter:
         SessionCursorAdapter sessionsAdapter = new SessionCursorAdapter(C, cursorSessionsResponse);
@@ -60,27 +63,26 @@ public class StudentSessionsFragment extends Fragment {
         sessionHistoryListView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
             @Override
             public void onItemClick(AdapterView<?> adapterView, View view, int position, long id) {
-                // Get tutor id
+                // Get session id
                 cursorSessionsResponse.moveToPosition(position);
                 int SESSION_ID = cursorSessionsResponse.getInt(cursorSessionsResponse.getColumnIndex("id"));
-                int TUTOR_ID = cursorSessionsResponse.getInt(cursorSessionsResponse.getColumnIndex("_id"));
 
-                // Add USER_ID and TUTOR_ID to session details fragment for displaying
+                // Add USER_ID and SESSION_ID to session details fragment for displaying
                 Bundle bundle = new Bundle();
                 bundle.putInt("USER_ID", USER_ID);
                 bundle.putInt("SESSION_ID", SESSION_ID);
-                bundle.putInt("TUTOR_ID", TUTOR_ID);
 
                 // Swap into new fragment
-                StudentHistoryDetailsFragment session = new StudentHistoryDetailsFragment();
+                TutorHistoryDetailsFragment session = new TutorHistoryDetailsFragment();
                 session.setArguments(bundle);
                 FragmentTransaction transaction = getFragmentManager().beginTransaction();
-                transaction.replace(R.id.student_fragment_container, session);
+                transaction.replace(R.id.tutor_fragment_container, session);
                 transaction.addToBackStack(null);
                 transaction.commit();
 
             }
         });
+
         return view;
     }
 }
