+++ app/src/main/res/layout/student_home_upc_session_details.xml
