+++ app/src/main/res/layout/student_sessions.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:gravity="center"
-    tools:context="cs.dal.krush.studentFragments.StudentSessionsFragment">
+    tools:context="cs.dal.krush.studentFragments.StudentHistoryFragment">
 
     <TextView
         android:id="@+id/sessionHistoryTitle"
